+++ taf/api/repository.py
@@ -28,8 +28,9 @@ from taf.log import taf_logger
 @log_on_end(DEBUG, "Finished adding or updating dependency", logger=taf_logger)
 @log_on_error(
     ERROR,
-    "An error occurred while adding a new dependency: {e!r}",
+    "An error occurred while adding a new dependency: {e}",
     logger=taf_logger,
+    on_exceptions=TAFError,
     reraise=True,
 )
 @check_if_clean
@@ -43,6 +44,7 @@ def add_dependency(
     library_dir: str = None,
     scheme: str = DEFAULT_RSA_SIGNATURE_SCHEME,
     custom=None,
+    prompt_for_keys=False,
 ):
     """
     Add a dependency (an authentication repository) to dependencies.json or update it if it was already added to this file.
@@ -132,22 +134,26 @@ def add_dependency(
         keystore,
         write=False,
         scheme=scheme,
+        prompt_for_keys=prompt_for_keys,
     )
 
     # update snapshot and timestamp calls write_all, so targets updates will be saved too
-    update_snapshot_and_timestamp(auth_repo, keystore, scheme=scheme)
+    update_snapshot_and_timestamp(
+        auth_repo, keystore, scheme=scheme, prompt_for_keys=prompt_for_keys
+    )
     commit_message = input("\nEnter commit message and press ENTER\n\n")
     auth_repo.commit(commit_message)
 
 
 @log_on_start(
-    INFO, "Creating a new authentication repository {repo_path:s}", logger=taf_logger
+    INFO, "Creating a new authentication repository {path:s}", logger=taf_logger
 )
 @log_on_end(INFO, "Finished creating a new repository", logger=taf_logger)
 @log_on_error(
     ERROR,
-    "An error occurred while creating a new repository: {e!r}",
+    "An error occurred while creating a new repository: {e}",
     logger=taf_logger,
+    on_exceptions=TAFError,
     reraise=True,
 )
 def create_repository(
@@ -306,8 +312,9 @@ def _determine_out_of_band_data(dependency, branch_name, out_of_band_commit):
 @log_on_end(DEBUG, "Finished removing dependency", logger=taf_logger)
 @log_on_error(
     ERROR,
-    "An error occurred while removing a dependency: {e!r}",
+    "An error occurred while removing a dependency: {e}",
     logger=taf_logger,
+    on_exceptions=TAFError,
     reraise=True,
 )
 @check_if_clean
@@ -316,6 +323,7 @@ def remove_dependency(
     dependency_name: str,
     keystore: str,
     scheme: str = DEFAULT_RSA_SIGNATURE_SCHEME,
+    prompt_for_keys: bool = False,
 ):
     """
     Remove a dependency (an authentication repository) from dependencies.json
@@ -370,10 +378,13 @@ def remove_dependency(
         keystore,
         write=False,
         scheme=scheme,
+        prompt_for_keys=prompt_for_keys,
     )
 
     # update snapshot and timestamp calls write_all, so targets updates will be saved too
-    update_snapshot_and_timestamp(auth_repo, keystore, scheme=scheme)
+    update_snapshot_and_timestamp(
+        auth_repo, keystore, scheme=scheme, prompt_for_keys=prompt_for_keys
+    )
     commit_message = input("\nEnter commit message and press ENTER\n\n")
     auth_repo.commit(commit_message)